Fix Always Hash Bug
This fixes a bug that causes Teleporter to always hash a local file
if the `-o` flag is given, even if the file does not exist on the
remote server. Now if a file does not exist, even though it is being
hashed on the client side in a thread, Teleporter will ignore the
hash and begin sending the file without waiting for the hash to
complete.