Add Delta transfer disable
Replace all network reads and writes with helper functions to
automatically add header information which includes the length
of the packet.

Removed all <= v0.5.5 structs and replaced them with new and more
versatile structs. This includes a features field which can be
expanded in future releases.

Cleaned up unused and/or unnecessary functions.