Implement XXH3 for the delta and file hash
This change will require changing the protocol fields, but cleans
up the serializeation/deserialization functions. It also speeds
up the hashing and comparisons due to it being a 64bit hash instead
of a 32byte hash. According to:
https://github.com/Cyan4973/xxHash/wiki/Collision-ratio-comparison
and
https://github.com/Cyan4973/xxHash/wiki/Performance-comparison
the XXH3 hash is extremely fast and has a low collision rate.

Also increase the number of chunks that we support from up to 150
per file to 2048.